backend/requirements.txt
# Core backend dependencies for Recipe Planner
fastapi
uvicorn
pymongo
python-dotenv
python-jose
passlib
bcrypt
apscheduler
requests
pytz
openai
pytest